refactor: adopt new classes, remove obsolete
