--- src/Yaapii.Atoms/Lists/ListOf.cs
@@ -1,69 +1,0 @@
-﻿// MIT License
-//
-// Copyright(c) 2019 ICARUS Consulting GmbH
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-using System;
-using System.Collections.Generic;
-using Yaapii.Atoms.Enumerable;
-
-namespace Yaapii.Atoms.Lists
-{
-    /// <summary>
-    /// Makes a readonly list.
-    /// </summary>
-    /// <typeparam name="T">type of items</typeparam>
-    [Obsolete("This class is obsolete and will be removed in future versions. Use List.Of or List.Live")]
-    public sealed class ListOf<T> : List.Envelope<T>
-    {
-        /// <summary>
-        /// ctor
-        /// </summary>
-        /// <param name="array">source array</param>
-        public ListOf(params T[] array) : this(new Many.Live<T>(array))
-        { }
-
-        /// <summary>
-        /// ctor
-        /// </summary>
-        /// <param name="src">source enumerator</param>
-        public ListOf(IEnumerator<T> src) : this(new Many.Of<T>(() => src))
-        { }
-
-        /// <summary>
-        /// ctor
-        /// </summary>
-        /// <param name="src">source enumerable</param>
-        public ListOf(IEnumerable<T> src) : base(() =>
-            {
-                var temp = new List<T>();
-                foreach (T item in src)
-                {
-                    temp.Add(item);
-                }
-
-                return temp;
-            },
-            true
-        )
-        { }
-
-    }
-}
