--- tests/Yaapii.Atoms.Tests/Lists/StickyTest.cs
@@ -1,46 +1,0 @@
-﻿// MIT License
-//
-// Copyright(c) 2019 ICARUS Consulting GmbH
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-using System.Threading;
-using Xunit;
-using Yaapii.Atoms.Scalar;
-
-namespace Yaapii.Atoms.Lists.Tests
-{
-    public sealed class StickyTest
-    {
-        [Fact]
-        public void IgnoresChangesInList()
-        {
-            int size = 2;
-            var list =
-                new StickyList<int>(
-                    new List.Live<int>(
-                        new Yaapii.Atoms.Enumerable.HeadOf<int>(
-                            new Yaapii.Atoms.Enumerable.Endless<int>(1),
-                            new ScalarOf<int>(() => Interlocked.Increment(ref size))
-                )));
-
-            Assert.Equal(new Enumerable.LengthOf(list).Value(), new Enumerable.LengthOf(list).Value());
-        }
-    }
-}
